pyproject.toml
[tool.poetry]
name = "llm-benchmark"
version = "0.1.0"
description = ""
authors = ["AleHD <alejandro.hernandezcano@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dacite = "^1.8.1"
toml = "^0.10.2"
jinja2 = "^3.1.4"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
seaborn = "^0.13.2"
wandb = "^0.17.0"
hydra-core = "^1.3.2"
plotly = "^5.24.0"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
